feat - App now caches the fetched articles and serves them if there is a connection error
